# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Xpart App OpenApi API Documentation",
        "description": "Xpart App Using L5 Swagger OpenApi description",
        "contact": {
            "email": "schneidershades@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/admin/withdrawals": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "allWithdrawals",
                "description": "allWithdrawals",
                "operationId": "allWithdrawals",
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/withdrawals/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "showWithdrawals",
                "description": "showWithdrawals",
                "operationId": "showWithdrawals",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "showWithdrawals Receipt number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "updateWithdrawals",
                "description": "updateWithdrawals",
                "operationId": "updateWithdrawals",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "updateWithdrawals Receipt Number ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdminWithdrawalsUpdateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/withdrawals/finalize": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "finalizeWithdrawals",
                "description": "finalizeWithdrawals",
                "operationId": "finalizeWithdrawals",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdminFinalizeTransferFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/withdrawals/verify/{receipt_number}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "verifyWithdrawals",
                "description": "verifyWithdrawals",
                "operationId": "verifyWithdrawals",
                "parameters": [
                    {
                        "name": "receipt_number",
                        "in": "path",
                        "description": "verifyWithdrawals Receipt Number ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AdminFinalizeTransferFormRequest": {
                "title": "Withdrawal Finalize Form Request Fields",
                "description": "Withdrawal Finalize Form request body data",
                "required": [
                    "email"
                ],
                "properties": {
                    "receipt_number": {
                        "title": "Order receipt payment reference",
                        "description": "Order  receipt payment reference",
                        "type": "string",
                        "example": "quote"
                    },
                    "otp": {
                        "title": "Order payment paystack otp",
                        "description": "Order payment paystack otp",
                        "type": "integer",
                        "example": "593349"
                    }
                },
                "type": "object"
            },
...
            "AdminWithdrawalsUpdateFormRequest": {
                "title": "Admin Withdrawals Update Form Request Fields",
                "description": "Admin Withdrawals Update request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "status",
                        "type": "string",
                        "example": "approve/decline"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}